requirements.txt
flask
flask_restful
gunicorn
newspaper3k
